[tuan] refactor for constants
