--- src/constante/index.js
@@ -1,7 +1,0 @@
-const { messageResponse } = require("./auth.constante")
-const productConstante = require("./product.constante")
-
-module.exports = {
-  messageResponse,
-  productConstante
-}
