+++ src/constants/index.js
@@ -1,0 +1,7 @@
+const { messageResponse } = require("./auth.constant")
+const productConstante = require("./product.constant")
+
+module.exports = {
+  messageResponse,
+  productConstante
+}
