+++ src/entities/report.entity.js
@@ -1,5 +1,5 @@
 const { EntitySchema } = require("typeorm")
-const { reportString, storeString } = require("../constante/entityName.js")
+const { reportString, storeString } = require("../constants/entityName.js")
 
 const reportEntity = new EntitySchema({
   name: reportString,
